Adding the project: Sliding Mode Controller and PID applied to a bldc motor
This one contains the pdf presentation and MATLAB/Simulink codes of one of the projects I did in the first year of my master's degree at Kirikkale University in the Electrical Machines Control course. The aim of the project was to design a Sliding Mode Controller (SMC) for a brushless-direct-current (BLDC) motor. Then perform comparison with a Proportional-Integral-Derivative (PID) Controller.